requirements.txt
PyYAML>=6.0
Pillow>=9.0.0
piexif>=1.1.3
python-whois>=0.7.3
dnspython>=2.0.0
requests>=2.25.1
python-magic>=0.4.27